Assignment_2/NAGphormer/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
appdirs=1.4.4=pyhd3eb1b0_0
blas=1.0=mkl
bottleneck=1.3.5=py38h7deecbd_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.0=py38h7f8727e_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=38.0.4=py38h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
dgl-cuda10.2=0.9.1post1=py38_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
idna=3.4=py38h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.1=py38h06a4308_0
jpeg=9e=h7f8727e_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20221030=h5eee18b_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_1
libunistring=0.9.10=h27cfd23_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
littleutils=0.2.2=py_0
lz4-c=1.9.4=h6a678d5_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py38h06a4308_0
numexpr=2.8.4=py38he184ba9_0
numpy=1.23.5=py38h14f4228_0
numpy-base=1.23.5=py38h31eccc5_0
ogb=1.3.5=pyhd8ed1ab_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1l=h7f98852_0
outdated=0.2.2=pyhd8ed1ab_0
packaging=22.0=py38h06a4308_0
pandas=1.5.2=py38h417a72b_0
pillow=9.3.0=py38h6a678d5_2
pip=22.3.1=py38h06a4308_0
pooch=1.4.0=pyhd3eb1b0_0
psutil=5.9.0=py38h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.0=h0371630_2
python-dateutil=2.8.2=pyhd8ed1ab_0
pytorch=1.11.0=py3.8_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pyhd8ed1ab_0
readline=7.0=h7b6447c_5
requests=2.28.1=py38h06a4308_0
scikit-learn=1.2.0=py38h6a678d5_0
scipy=1.10.0=py38h14f4228_0
setuptools=65.6.3=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.33.0=h62c20be_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.11.0=py38_cu102
torchvision=0.12.0=py38_cu102
tqdm=4.64.1=py38h06a4308_0
typing_extensions=4.4.0=py38h06a4308_0
urllib3=1.26.14=py38h06a4308_0
wheel=0.38.4=py38h06a4308_0
xz=5.2.10=h5eee18b_1
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
